create component for posters and add loading state while posters are being fetched
